SAA-1528 API endpoint GET /subject-access-request/prisoner/{prisonerNumber}
PR comments. Removed custom class for the attendance summary and mapped in the service layer. Moved the repository reference to the SarRepository

Add attendances for the subject access request. Uses the v_all_attendance view and create a summary of the attendances in the response for a given prisoner number, from and to date.
Session date is used for the searching of from and to date
e.g.

"attendanceSummary": [
        {
            "attendanceReasonCode": "ATTENDED",
            "count": 2
        },
        {
            "attendanceReasonCode": "SUSPENDED",
            "count": 1
        }
    ],