SAA-1528 API endpoint GET /subject-access-request/prisoner/{prisonerNumber}
PR comments. Only return non-null values for the attendance summary grouping

Add attendances for the subject access request. Uses the v_all_attendance view and create a summary of the attendances in the response for a given prisoner number, from and to date.
Session date is used for the searching of from and to date
e.g.

"attendanceSummary": [
        {
            "attendanceReasonCode": "ATTENDED",
            "count": 2
        },
        {
            "attendanceReasonCode": "SUSPENDED",
            "count": 1
        }
    ],