BAVL-1030 attendees should not be removed from court and probation video appointments
This change stops attendees from being removed from video link court and probation appointments. These appointment types are (now) managed outside of A&A via the book a video link service.

For example, when a prisoner is released the appointment is cancelled by the book a video link service making a call the A&A API.  A&A should not take further any action for these appointment types on the release of prisoners.